updating main.go
