+++ apis/main/main.go
@@ -8,15 +8,11 @@ import (
 	"net/http"
 )
 
-// var (
-// 	users    = make(map[string]user.User)
-// 	housings = make(map[string]housing.Housing)
-// )
-
 func main() {
 	fmt.Println("Running")
-	http.HandleFunc("/register", registerHandler)
-	http.HandleFunc("/users", getUsersHandler)
+	http.HandleFunc("/api/register", registerHandler)
+	http.HandleFunc("/api/users", getUsersHandler)
+	http.HandleFunc("/api/login", loginHandler)
 	http.ListenAndServe(":8080", nil)
 	fmt.Println("Server up")
 }
@@ -41,6 +37,7 @@ func registerHandler(w http.ResponseWriter, r *http.Request) {
 	data.Users[lastname] = user.User{FName: firstname, LName: lastname, UserName: username, Password: password}
 
 	fmt.Fprintf(w, "Registration successful")
+	w.WriteHeader(http.StatusOK)
 }
 
 func getUsersHandler(w http.ResponseWriter, r *http.Request) {
@@ -52,4 +49,24 @@ func getUsersHandler(w http.ResponseWriter, r *http.Request) {
 	w.Header().Set("Content-Type", "application/json")
 	json.NewEncoder(w).Encode(data.Users)
 
+}
+
+func loginHandler(w http.ResponseWriter, r *http.Request) {
+	if r.Method != http.MethodGet {
+		http.Error(w, "Method not allowed", http.StatusMethodNotAllowed)
+		return
+	}
+
+	if verifyLogin(r.FormValue("username"), r.FormValue("passowrd")) {
+
+		fmt.Fprintf(w, "Login successful")
+		w.WriteHeader(http.StatusOK)
+	} else {
+		http.Error(w, "Invalid username or password", http.StatusUnauthorized)
+		return
+	}
+}
+
+func verifyLogin(username string, password string) bool {
+	return true
 }
